backend/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
asyncpg==0.30.0
alembic==1.14.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Settings
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# AWS S3
boto3==1.35.86

# Utilities
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
